+++ src/main/java/com/zaknein/the_cowboyday_API/repository/TontosRepository.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 import com.zaknein.the_cowboyday_API.entity.Tontos;
 
 @Repository
-public interface TontosRepository extends JpaRepository<Tontos, Long> {
+public interface TontosRepository extends JpaRepository<Tontos, Integer> {
