+++ src/main/java/com/zaknein/the_cowboyday_API/service/TitulosService.java
@@ -35,7 +35,7 @@ public class TitulosService {
         return titulosRepository.save(titulos);
     }
 
-    public void deleteTitulos(Long id) {
+    public void deleteTitulos(int id) {
 
         Titulos tituloToDelete = titulosRepository.findById(id)
                 .orElseThrow(()-> new RuntimeException("no titulo con ese id"));
